+++ Files/layers.py
@@ -1,9 +1,13 @@
 # layers.py
 
+import json
+
 from Files.neuron import Neuron
 
 
 class Layers:
+    _neurons_path: str = "Files/neurons.json"
+    
     def __init__(self, config: dict, data: dict):
         self.config: dict = config
         self.data: dict = data
@@ -20,10 +24,28 @@ class Layers:
                     bias=biases[l][n],
                     activation=activations[l]
                 )
-                for n in range(l)
+                for n in range(layer)
             ]
-            for l in len(layers)
+            for l, layer in enumerate(layers)
         ]
+
+    
+    def save(self) -> None:
+        data: dict = {}
+        for layer in self.layers:
+            for n in layer:
+                data[n.name] = {
+                    "weights": n.weights,
+                    "bias": n.bias,
+                    "score": n.score,
+                    "activation": n.activation,
+                    "slope": n.slope,
+                    "delta": n.delta,
+                    "input": n.input,
+                    "output": n.output
+                }
+        with open(self._neurons_path, "w") as file:
+            json.dump(data, file, indent=4)
 
     
     def __iter__(self):